requirements.txt
streamlit
langchain
openai
pypdf2
faiss-cpu
python-dotenv
streamlit-extras